requirements.txt
tensorflow>=2.8.0
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.1
scikit-learn>=1.0.2
prometheus-client>=0.14.1
kubernetes>=23.6.0
click>=8.0.0
pytest>=7.0.0
pylint>=2.12.0
bandit>=1.7.0
safety>=2.3.1